fixes : pr #1 | proxy client update | ethereum instance
