+++ crates/starknet-proxy-client/Cargo.toml
@@ -15,3 +15,4 @@ num-traits = { workspace = true }
 async-trait = { workspace = true }
 serde_json = { workspace = true }
 utils = { path = "../utils" }
+ethereum-instance = { path = "../ethereum-instance" }
